app/services/crawler_service.py
import os
import requests
from typing import Optional, Dict, Any, List
from bs4 import BeautifulSoup
import re
import asyncio

from app.services.ai_service import get_ai_response

# 网页爬取的通用头信息
HEADERS = {
    "User-Agent": "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/91.0.4472.124 Safari/537.36",
    "Accept": "text/html,application/xhtml+xml,application/xml;q=0.9,image/webp,*/*;q=0.8",
    "Accept-Language": "zh-CN,zh;q=0.9,en;q=0.8",
}

async def crawl_article(url: str) -> str:
    """
    爬取网页文章内容
    
    Args:
        url: 网页URL
        
    Returns:
        提取的文章内容
    """
   


async def _crawl_bilibili(url: str) -> str:
    """Bilibili视频爬取 (简化版)"""
    
async def summarize_content(
    content: str,
    extract_type: str = "summary",
    max_length: Optional[int] = None
) -> str:
    """
    使用AI对爬取的内容进行处理
    
    Args:
        content: 原始内容
        extract_type: 提取类型 (summary, outline)
        max_length: 最大长度限制
        
    Returns:
        处理后的内容
    """
   