requirements.txt
fastapi>=0.68.0,<0.69.0
uvicorn>=0.15.0,<0.16.0
pydantic>=1.8.0,<2.0.0
python-dotenv>=0.19.0,<0.20.0
requests>=2.26.0,<2.27.0
beautifulsoup4>=4.10.0,<4.11.0
openai>=0.27.0
langchain>=0.0.200
python-multipart>=0.0.5
